move `Acker` into separate file
