+++ lib/vector-core/src/buffers/mod.rs
@@ -1,18 +1,16 @@
+mod acker;
+#[cfg(feature = "disk-buffer")]
+pub mod disk;
+
 use crate::event::Event;
-use futures::{channel::mpsc, task::AtomicWaker, Sink, SinkExt};
+pub use acker::Acker;
+use futures::{channel::mpsc, Sink, SinkExt};
 use pin_project::pin_project;
 use serde::{Deserialize, Serialize};
 use std::{
     pin::Pin,
-    sync::{
-        atomic::{AtomicUsize, Ordering},
-        Arc,
-    },
     task::{Context, Poll},
 };
-
-#[cfg(feature = "disk-buffer")]
-pub mod disk;
 
 #[derive(Deserialize, Serialize, Debug, PartialEq, Copy, Clone)]
 #[serde(rename_all = "snake_case")]
@@ -62,44 +60,6 @@ impl BufferInputCloner {
                 }
             }
         }
-    }
-}
-
-#[derive(Debug, Clone)]
-pub enum Acker {
-    Disk(Arc<AtomicUsize>, Arc<AtomicWaker>),
-    Null,
-}
-
-impl Acker {
-    // This method should be called by a sink to indicate that it has
-    // successfully flushed the next `num` events from its input stream. If
-    // there are events that have flushed, but events that came before them in
-    // the stream have not been flushed, the later events must _not_ be acked
-    // until all preceding elements are also acked.  This is primary used by the
-    // on-disk buffer to know which events are okay to delete from disk.
-    pub fn ack(&self, num: usize) {
-        // Only ack items if the amount to ack is larger than zero.
-        if num > 0 {
-            match self {
-                Acker::Null => {}
-                Acker::Disk(counter, notifier) => {
-                    counter.fetch_add(num, Ordering::Relaxed);
-                    notifier.wake();
-                }
-            }
-
-            // TODO re-enable
-            // emit!(EventOut { count: num });
-        }
-    }
-
-    pub fn new_for_testing() -> (Self, Arc<AtomicUsize>) {
-        let ack_counter = Arc::new(AtomicUsize::new(0));
-        let notifier = Arc::new(AtomicWaker::new());
-        let acker = Acker::Disk(Arc::clone(&ack_counter), Arc::clone(&notifier));
-
-        (acker, ack_counter)
     }
 }
